[Rename] StartGame -> startGames 폴더명 변경
- git에서 대소문자 인식이 안돼서, 폴더명에 s를 추가했습니다.
- 추후 startGame로 다시 바꾸면 좋을 듯합니다.
