+++ components/screen/startGames/MiniNetworking.tsx
