+++ components/screen/startGames/ThankCircle.tsx
